apps/server/pyproject.toml
[project]
name = "ogen-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "ogen-stream", # 이름 참조
]

[tool.uv.sources]
ogen-stream = { workspace = true }